+++ src/main/java/net/fauxpark/oled/impl/SSD1306SPIImpl.java
@@ -49,85 +49,6 @@ public class SSD1306SPIImpl extends SSD1306Display {
 		spi = SpiFactory.getInstance(channel, SPI_SPEED);
 	}
 
-
-
-
-
-
-	@Override
-	public void setInverted(boolean inverted) {
-		command(inverted ? Command.INVERT_DISPLAY : Command.NORMAL_DISPLAY);
-		super.setInverted(inverted);
-	}
-
-	@Override
-	public void setContrast(int contrast) {
-		if(contrast < 0 || contrast > 255) {
-			return;
-		}
-
-		command(Command.SET_CONTRAST, contrast);
-		super.setContrast(contrast);
-	}
-
-	@Override
-	public void setOffset(int offset) {
-		command(Command.SET_DISPLAY_OFFSET, offset);
-		super.setOffset(offset);
-	}
-
-	@Override
-	public void setHFlipped(boolean hFlipped) {
-		if(hFlipped) {
-			command(Command.SET_SEGMENT_REMAP);
-		} else {
-			command(Command.SET_SEGMENT_REMAP_REVERSE);
-		}
-
-		// Horizontal flipping is not immediate
-		display();
-		super.setHFlipped(hFlipped);
-	}
-
-	@Override
-	public void setVFlipped(boolean vFlipped) {
-		if(vFlipped) {
-			command(Command.SET_COM_SCAN_INC);
-		} else {
-			command(Command.SET_COM_SCAN_DEC);
-		}
-
-		super.setVFlipped(vFlipped);
-	}
-
-	@Override
-	public void scrollHorizontally(boolean direction, int start, int end, int speed) {
-		command(direction ? Command.LEFT_HORIZONTAL_SCROLL : Command.RIGHT_HORIZONTAL_SCROLL, Constant.DUMMY_BYTE_00, start, speed, end, Constant.DUMMY_BYTE_00, Constant.DUMMY_BYTE_FF);
-	}
-
-	@Override
-	public void scrollDiagonally(boolean direction, int start, int end, int offset, int rows, int speed, int step) {
-		command(Command.SET_VERTICAL_SCROLL_AREA, offset, rows);
-		command(direction ? Command.VERTICAL_AND_LEFT_HORIZONTAL_SCROLL : Command.VERTICAL_AND_RIGHT_HORIZONTAL_SCROLL, Constant.DUMMY_BYTE_00, start, speed, end, step);
-	}
-
-	@Override
-	public void startScroll() {
-		command(Command.ACTIVATE_SCROLL);
-		super.startScroll();
-	}
-
-	@Override
-	public void stopScroll() {
-		command(Command.DEACTIVATE_SCROLL);
-		super.stopScroll();
-	}
-
-	@Override
-	public void noOp() {
-		command(Command.NOOP);
-	}
-
 	@Override
 	public void command(int command, int... params) {
 		dcPin.setState(false);
